+++ biofeedback_cube/fx/cat_jam.py
@@ -19,7 +19,7 @@ def open_images():
     with imageio.read(path) as f:
         for im in f.iter_data():
             flipped_im = im[::-1, ::-1, :]
-            images.append(flipped_im.astype(np.float64) / 255)
+            images.append(flipped_im.astype(np.float32) / 255)
         meta['length'] = f.get_length()
         meta['duration'] = f.get_meta_data()['duration']
         meta['total_time'] = (meta['length'] * meta['duration']) / 1000
